Add new parser logic for workflows
