+++ src/ert/parsing/config_keywords.py
@@ -1,17 +1,125 @@
-from enum import Enum, auto
+from enum import Enum
 
 
-class SchemaType(Enum):
-    CONFIG_STRING = auto()
-    CONFIG_INT = auto()
-    CONFIG_FLOAT = auto()
-    CONFIG_PATH = auto()
-    CONFIG_EXISTING_PATH = auto()
-    CONFIG_BOOL = auto()
-    CONFIG_CONFIG = auto()
-    CONFIG_BYTESIZE = auto()
-    CONFIG_EXECUTABLE = auto()
-    CONFIG_ISODATE = auto()
-    CONFIG_INVALID = auto()
-    CONFIG_RUNTIME_INT = auto()
-    CONFIG_RUNTIME_FILE = auto()
+class ConfigKeys(str, Enum):
+    # Note: The values of these must correspond to
+    # those used in the ert config, unless we make
+    # the keys match up and set logic checking
+    # config keys against strings (e.g., in check_required)
+    # to use only names
+    # These keys are used as options in KEY:VALUE statements
+    BASE_SURFACE_KEY = "BASE_SURFACE"
+    DEFINE_KEY = "DEFINE"
+    FORWARD_INIT_KEY = "FORWARD_INIT"
+    GENERAL_KEY = "GENERAL"
+    INCLUDE_KEY = "INCLUDE"
+    INIT_FILES_KEY = "INIT_FILES"
+    INIT_TRANSFORM_KEY = "INIT_TRANSFORM"
+    INPUT_FORMAT_KEY = "INPUT_FORMAT"
+    INPUT_TRANSFORM_KEY = "INPUT_TRANSFORM"
+    MAX_KEY = "MAX"
+    MIN_KEY = "MIN"
+    OUTPUT_FILE_KEY = "OUTPUT_FILE"
+    OUTPUT_TRANSFORM_KEY = "OUTPUT_TRANSFORM"
+    PARAMETER_KEY = "PARAMETER"
+    REPORT_STEPS_KEY = "REPORT_STEPS"
+    RESULT_FILE_KEY = "RESULT_FILE"
+    TEMPLATE_KEY = "TEMPLATE"
+    ANALYSIS_COPY_KEY = "ANALYSIS_COPY"
+    ANALYSIS_SET_VAR_KEY = "ANALYSIS_SET_VAR"
+    ANALYSIS_SELECT_KEY = "ANALYSIS_SELECT"
+    DATA_ROOT_KEY = "DATA_ROOT"
+    DATA_FILE_KEY = "DATA_FILE"
+    DATA_KW_KEY = "DATA_KW"
+    ECLBASE_KEY = "ECLBASE"
+    ENKF_ALPHA_KEY = "ENKF_ALPHA"
+    ENSPATH_KEY = "ENSPATH"
+    ITER_CASE_KEY = "ITER_CASE"
+    ITER_COUNT_KEY = "ITER_COUNT"
+    ITER_RETRY_COUNT_KEY = "ITER_RETRY_COUNT"
+    FIELD_KEY = "FIELD"
+    FORWARD_MODEL_KEY = "FORWARD_MODEL"
+    GEN_DATA_KEY = "GEN_DATA"
+    GEN_KW_KEY = "GEN_KW"
+    GEN_KW_TAG_FORMAT_KEY = "GEN_KW_TAG_FORMAT"
+    GEN_KW_EXPORT_NAME_KEY = "GEN_KW_EXPORT_NAME"
+    GRID_KEY = "GRID"
+    HISTORY_SOURCE_KEY = "HISTORY_SOURCE"
+    INSTALL_JOB_KEY = "INSTALL_JOB"
+    INSTALL_JOB_DIRECTORY_KEY = "INSTALL_JOB_DIRECTORY"
+    JOB_SCRIPT_KEY = "JOB_SCRIPT"
+    JOBNAME_KEY = "JOBNAME"
+    LICENSE_PATH_KEY = "LICENSE_PATH"
+    MAX_SUBMIT_KEY = "MAX_SUBMIT"
+    NUM_REALIZATIONS_KEY = "NUM_REALIZATIONS"
+    MIN_REALIZATIONS_KEY = "MIN_REALIZATIONS"
+    OBS_CONFIG_KEY = "OBS_CONFIG"
+    QUEUE_SYSTEM = "QUEUE_SYSTEM"
+    QUEUE_OPTION_KEY = "QUEUE_OPTION"
+    HOOK_WORKFLOW_KEY = "HOOK_WORKFLOW"
+    REFCASE_KEY = "REFCASE"
+    RUNMODE_KEY = "RUNMODE"
+    RUNPATH_FILE_KEY = "RUNPATH_FILE"
+    RUNPATH_KEY = "RUNPATH"
+    RUN_TEMPLATE_KEY = "RUN_TEMPLATE"
+    SCHEDULE_PREDICTION_FILE_KEY = "SCHEDULE_PREDICTION_FILE"
+    SETENV_KEY = "SETENV"
+    SIMULATION_JOB_KEY = "SIMULATION_JOB"
+    STD_CUTOFF_KEY = "STD_CUTOFF"
+    SUMMARY_KEY = "SUMMARY"
+    SURFACE_KEY = "SURFACE"
+    UPDATE_LOG_PATH_KEY = "UPDATE_LOG_PATH"
+    UPDATE_PATH_KEY = "UPDATE_PATH"
+    RANDOM_SEED_KEY = "RANDOM_SEED"
+    WORKFLOW_JOB_DIRECTORY_KEY = "WORKFLOW_JOB_DIRECTORY"
+    LOAD_WORKFLOW_KEY = "LOAD_WORKFLOW"
+    LOAD_WORKFLOW_JOB_KEY = "LOAD_WORKFLOW_JOB"
+    STOP_LONG_RUNNING_KEY = "STOP_LONG_RUNNING"
+    MAX_RUNTIME_KEY = "MAX_RUNTIME"
+    TIME_MAP_KEY = "TIME_MAP"
+    UPDATE_SETTING_KEY = "UPDATE_SETTINGS"
+    NUM_CPU_KEY = "NUM_CPU"
+
+    CONFIG_DIRECTORY_KEY = "CONFIG_DIRECTORY"
+
+    SLURM_SBATCH_OPTION = "SBATCH"
+    SLURM_SCANCEL_OPTION = "SCANCEL"
+    SLURM_SCONTROL_OPTION = "SCONTROL"
+    SLURM_SQUEUE_OPTION = "SQUEUE"
+    SLURM_PARTITION_OPTION = "PARTITION"
+    SLURM_SQUEUE_TIMEOUT_OPTION = "SQUEUE_TIMEOUT"
+
+    # Observe that the SLURM_MAX_RUNTIME_OPTION expects a time limit in seconds,
+    # whereas slurm uses a time limit in minutes
+    SLURM_MAX_RUNTIME_OPTION = "MAX_RUNTIME"
+    SLURM_MEMORY_OPTION = "MEMORY"
+    SLURM_MEMORY_PER_CPU_OPTION = "MEMORY_PER_CPU"
+
+    # For the EXCLUDE and INCLUDE host options the slurm driver
+    # maintains an internal list of hostnames, and the option can be called
+    # repeatedly. It is possible to add multiple hosts separated by space or comma
+    # in one option call:
+    #
+    # QUEUE_OPTION SLURM EXCLUDE_HOST host1,host2,host3
+    # QUEUE_OPTION SLURM EXCLUDE_HOST host5 host6,host7
+    SLURM_EXCLUDE_HOST_OPTION = "EXCLUDE_HOST"
+    SLURM_INCLUDE_HOST_OPTION = "INCLUDE_HOST"
+
+
+class ConfigArgAtIndex(str, Enum):
+    pass
+
+
+class RunModes(ConfigArgAtIndex):
+    RUN_MODE_PRE_SIMULATION_NAME = "PRE_SIMULATION"
+    RUN_MODE_POST_SIMULATION_NAME = "POST_SIMULATION"
+    RUN_MODE_PRE_UPDATE_NAME = "PRE_UPDATE"
+    RUN_MODE_POST_UPDATE_NAME = "POST_UPDATE"
+    RUN_MODE_PRE_FIRST_UPDATE_NAME = "PRE_FIRST_UPDATE"
+
+
+class QueueOptions(ConfigArgAtIndex):
+    LSF = "LSF"
+    LOCAL = "LOCAL"
+    TORQUE = "TORQUE"
+    SLURM = "SLURM"
